Fix leaked connections when PFS queries contain null digests
When the events_statements_summary_by_digest table in the performance schema
reaches the performance_schema_digests_size, it will aggregate all rows in a
row will a null digest.

When this null value is returned and parsed by `rows.Scan()`, an error occurs,
and the method returns early. Since `rows` was never fully iterated, it does
not automatically close itself as it usually does.

This patch adds a `defer rows.Close()` statement to prevent this issue, and
adds some additional logging when it occurs.